pyproject.toml
[project]
name = "referal-system-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "black>=25.1.0",
    "fastapi>=0.115.11",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "python-jose>=3.4.0",
    "pytz>=2025.1",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.10.6",
    "aiofiles>=24.1.0",
    "pytest>=8.3.5",
    "jinja2>=3.1.6",
    "httpx>=0.28.1",
    "bcrypt==4.0.1",
]
